refactor(l5): move tools/search-brain and tools/fts-config from src/ into server/tools/legacy-search behind L5 adapters (issue 864)
src/tools/search-brain.ts was a 2100-line module holding the whole legacy
search_brain pipeline. It moves under server/tools/legacy-search/ and is split
by responsibility, so each sibling meets the server/ standard (functions under
100 code lines, complexity under 10, at most 4 params, files under 500 code
lines). src/tools/fts-config.ts moves the same way.

Split, by table group -> file -> code lines:

  legacy-search-tables-and-parsing.ts   304  readable tables, relational query
                                             parse, search-embedding generation
  legacy-search-rows-and-fallback.ts    290  row shapes, evidence, dedupe, the
                                             shared fallback merge
  legacy-search-query-builders.ts       534  CTE and query construction
  legacy-search-arms.ts                 339  relational, vector, FTS searches
  legacy-search-execute.ts              382  RRF merge, usage tracking,
                                             executeSearch
  legacy-search-fallback-executors.ts    82  the shared-namespace fallback
                                             executors
  legacy-search-register.ts             369  the MCP tool registration
  fts-config.ts                         258  the FTS language configuration
  search-brain.ts                        45  the assembled re-export surface

Options types replace the long positional lists on the internal builders
(rule M15): TableCTEOptions and FtsCTEOptions for the two CTE builders, and
TrackUsageOptions and ExecuteSearchOptions for the execute path. Each CTE
builder now dispatches to one helper per table family (entities,
session_events, and the standard brain tables), which is where the shape
actually diverges; the SQL text each emits is unchanged.

Reader contracts (rule M13a): server/ code may not read the environment, so the
search-embedding timeout and the FTS environment both arrive through registered
readers. The src/ adapters register readers over process.env that perform
exactly the parse the original modules did, at the same moment they used to
happen -- per call, so a test that mutates the environment after import still
sees its change. The default readers return undefined, so a caller that
registers nothing gets the original defaults and no server/ path reads
process.env.

src/tools/search-brain.ts and src/tools/fts-config.ts remain as L5 adapters
(rule M9): they restore the legacy positional call form over the options-object
server/ functions and re-export the names the legacy call sites use. The
adapter derives its tier type from the moved options type rather than importing
it from ../types.ts, so every relative specifier an adapter names resolves under
server/, which is what the done-means check's clause A requires.

server/main.ts is unchanged by this commit and has no staged diff: it names
neither module, so the move reaches it through no path it holds.

Verified: bunx tsc --noEmit -> 0; oxlint --deny-warnings over the moved tree and
both adapters -> 0, from 19 findings before; the seven named test files ->
171 pass, 0 fail; bun contracts/check-parity.ts -> passed; the done-means check
-> PASS; src-runtime-closure 46 -> 45.
