+++ src/tools/search-brain.ts
@@ -1,2105 +1,135 @@
-import { z } from "zod";
-import type { McpServer } from "@modelcontextprotocol/sdk/server/mcp.js";
-import { toSql } from "pgvector/pg";
-import { canRead } from "../permissions.ts";
-import { canReadNamespace, namespaceFilterFor } from "../read-policy.ts";
-import {
-  canonicalNamespace,
-  isSharedNamespace,
-  sharedNamespaceConfig,
-} from "../shared-namespace.ts";
-import {
-  appendSourceScopeParam,
-  sourceScopeAuthorizationError,
-  sourceScopeFilterSql,
-  sourceScopeSchema,
-  type SourceScope,
-} from "../source-refs.ts";
-import type { AuthInfo, LinkRelation, Role, Table, Tier } from "../types.ts";
-import type { ToolDeps } from "./index.ts";
-import { logger } from "../logger.ts";
-import { describeError } from "../observability/index.ts";
-import {
-  ALL_TABLES,
-  SOURCE_LABELS,
-  CONTENT_PREVIEW,
-  FTS_SOURCE_TEXT,
-  LINK_RELATIONS,
-  TABLE_ALIAS,
-  VALID_TIERS,
-} from "./table-constants.ts";
-import {
-  setActiveMcpTraceMetadata,
-  traceRetrievalSpan,
-  traceRetrievalSpanSync,
-} from "../../server/observability/langfuse-tracing.ts";
-import {
-  DEFAULT_FTS_CONFIG,
-  ftsConfigLiteral,
-  ftsStatementTimeoutMs,
-  requestFtsConfig,
-  SUPPORTED_FTS_CONFIGS,
-  type FtsConfig,
-} from "./fts-config.ts";
+// L5 adapter (issue 864): legacy call form over server/tools/legacy-search/search-brain.ts; retired with src/ at L6.
+//
+// The moved module takes its search-embedding timeout from a registered reader
+// (rule M13a) because server/ code may not read process.env. This adapter
+// registers a reader that performs exactly the parse the original module did,
+// at the same moment it used to happen -- per call, so a test that mutates the
+// environment after import still sees its change.
+import { setFtsEnvReader } from "../../server/tools/legacy-search/fts-config.ts";
+import { setSearchEmbeddingTimeoutReader } from "../../server/tools/legacy-search/search-brain.ts";
 
-export { ALL_TABLES };
+setFtsEnvReader(() => process.env);
 
-/** Reverse map: singular label -> table name for tracking UPDATEs */
-const LABEL_TO_TABLE: Record<string, Table> = Object.fromEntries(
-  Object.entries(SOURCE_LABELS).map(([table, label]) => [
-    label,
-    table as Table,
-  ]),
-) as Record<string, Table>;
-
-export type SearchMode = "hybrid" | "vector" | "keyword";
-type NamespaceFilter = string | string[];
-/**
- * Tables the retrieval arms can read.
- *
- * `Table` is the PHYSICAL-table type: it drives `PERMISSIONS`, the REST table
- * enum, and every write path, so widening it would demand a permission row and
- * a write contract for corpora that have neither. `entities` and
- * `session_events` are read-only retrieval sources instead — searchable, never
- * writable through this surface — and each carries its own CTE branch below
- * because its columns do not match the shared shape.
- */
-export type SearchTable = Table | "entities" | "session_events";
-
-/**
- * Session events are namespaced INDIRECTLY. `ob_session_events` has no
- * `namespace` column at all (verified against information_schema on the
- * dogfood DB, 2026-08-07); scope lives on `ob_session_lanes` and is reached
- * only through `ob_session_events.lane_id`. Every CTE below therefore joins
- * the lane table and applies the auth-derived namespace predicate to
- * `lane.namespace` — omitting that join does not merely widen results, it
- * exposes every agent's session history to every other namespace.
- *
- * The corpus also lacks `archived_at`, `tier`, `tags`, `usefulness_score`,
- * `access_count`, `search_vector`, and `source_refs`. Those slots are filled
- * with typed literals so the UNION arms stay type-compatible, and the
- * consequences are deliberate:
- *   - no stored `search_vector` -> the lexical arm matches with ILIKE on
- *     `content` rather than `to_tsvector`, so it is substring-based, not
- *     stemmed. That is the honest capability of a column that does not exist.
- *   - no `tier` -> `importance` carries the same hot/warm/cold vocabulary and
- *     the same CHECK constraint, so it is used as the tier for filtering and
- *     for TIER_BOOST.
- *   - no `source_refs` -> a source-scope filter cannot be satisfied, so the
- *     corpus is excluded when one is requested rather than silently ignoring
- *     the scope (which would leak out-of-scope rows into a scoped query).
- */
-const SESSION_EVENTS_SOURCE_LABEL = "session_event";
-
-/**
- * The recall sources a role may read, in one place.
- *
- * This exists because the selection was previously OPEN-CODED at each call
- * site, and that is exactly how #433 defect 1 happened: `search_brain` added
- * `entities` to its list, `brain_answer` did not, and nothing added
- * `session_events` anywhere -- so a corpus of 11,136 rows was unreachable from
- * the tool agents actually ask. Three copies of a list is three chances to
- * forget one. Callers that need the default set MUST use this function so a
- * new source is added once and appears everywhere at once.
- *
- * `entities` and `session_events` have no PERMISSIONS row: they are read-only
- * retrieval sources, never written through this surface, and both ride the
- * `sessions` read permission.
- */
-export function readableSearchTables(
-  role: Role,
-  options: { includeEntities?: boolean } = {},
-): SearchTable[] {
-  const tables: SearchTable[] = ALL_TABLES.filter((table) =>
-    canRead(role, table),
-  );
-  if (canRead(role, "sessions")) {
-    // `entities` is opt-in because the two recall surfaces legitimately differ:
-    // search_brain returns rows for a caller to read, while brain_answer cites
-    // extractive evidence, and an entity row's preview is a name label rather
-    // than a statement. Silently adding it to brain_answer would be a behavior
-    // change #433 never asked for. `session_events` is unconditional: it is
-    // real recorded content and its absence IS the defect.
-    if (options.includeEntities) tables.push("entities");
-    tables.push("session_events");
-  }
-  return tables;
-}
-export type { SourceScope };
-
-type ExecuteSearchOptions = {
-  enableGraph?: boolean;
-  /**
-   * Text-search configuration for the lexical arm. When unset, defaults to the
-   * shared english configuration. The public `search_brain` handler is the only
-   * boundary that resolves its request argument and deployment env; sibling
-   * executeSearch callers remain byte-compatible unless they explicitly pass a
-   * config.
-   */
-  ftsConfig?: FtsConfig;
-};
-
-const NON_ENGLISH_FTS_AUTHORIZATION_ERROR =
-  "Permission denied: non-English FTS configuration requires admin or ob-admin";
-
-/** RRF constant -- standard value from Cormack et al. 2009 */
-const RRF_K = 60;
-
-/** Over-fetch multiplier for hybrid mode (fetch N*3 from each path, merge to N) */
-const HYBRID_FETCH_MULTIPLIER = 3;
-const DEFAULT_SEARCH_EMBEDDING_TIMEOUT_MS = 3000;
-const RELATIONAL_GRAPH_FETCH_LIMIT = 50;
-
-/** Tier-based RRF score adjustments for cognitive tiering */
-export const TIER_BOOST: Record<Tier, number> = {
-  hot: 0.3,
-  warm: 0,
-  cold: -0.2,
-};
-
-/** Scoring weights for vector search ranking formula */
-const VECTOR_WEIGHT = 0.7;
-const USEFULNESS_WEIGHT = 0.15;
-const AGE_WEIGHT = 0.0001;
-
-/** Per-table importance weights: primary content > summaries */
-const TABLE_WEIGHT: Record<string, number> = {
-  thought: 1.2,
-  decision: 1.2,
-  relationship: 1.0,
-  project: 0.9,
-  session: 0.8,
-  entity: 1.0,
-};
-
-type RelationalDirection = "incoming" | "outgoing";
-
-const RELATIONAL_INCOMING_QUERY_PATTERN =
-  /^what\s+(?:is\s+)?(?:was\s+)?(?<relation>depends on|blocked by|implemented by|decided by|supersedes|duplicates|contradicts|mentions|relates to)\s+(?<seed>[^?]{1,160})\??$/iu;
-
-const RELATIONAL_OUTGOING_DEPENDS_PATTERN =
-  /^what\s+does\s+(?<seed>[^?]{1,160})\s+depend\s+on\??$/iu;
-
-const RELATIONAL_OUTGOING_BLOCKED_PATTERN =
-  /^what\s+(?:is\s+)?(?<seed>[^?]{1,160})\s+blocked\s+by\??$/iu;
-
-const RELATION_ALIASES: Record<string, LinkRelation> = {
-  "depends on": "depends_on",
-  "blocked by": "blocked_by",
-  "implemented by": "implemented_by",
-  "decided by": "decided_by",
-  supersedes: "supersedes",
-  duplicates: "duplicates",
-  contradicts: "contradicts",
-  mentions: "mentions",
-  "relates to": "relates_to",
-};
-
-type RelationalQuery = {
-  relation: LinkRelation;
-  seed: string;
-  direction: RelationalDirection;
-};
-
-function parseRelationalQuery(query: string): RelationalQuery | undefined {
-  const trimmed = query.trim();
-  const outgoingDepends =
-    RELATIONAL_OUTGOING_DEPENDS_PATTERN.exec(trimmed)?.groups;
-  if (outgoingDepends?.seed) {
-    const seed = outgoingDepends.seed.trim().replace(/\s+/g, " ");
-    if (!seed) return undefined;
-    return {
-      relation: "depends_on",
-      seed,
-      direction: "outgoing",
-    };
-  }
-  const outgoingBlocked =
-    RELATIONAL_OUTGOING_BLOCKED_PATTERN.exec(trimmed)?.groups;
-  if (outgoingBlocked?.seed) {
-    const seed = outgoingBlocked.seed.trim().replace(/\s+/g, " ");
-    if (!seed) return undefined;
-    return {
-      relation: "blocked_by",
-      seed,
-      direction: "outgoing",
-    };
-  }
-
-  const groups = RELATIONAL_INCOMING_QUERY_PATTERN.exec(trimmed)?.groups;
-  if (!groups?.relation || !groups.seed) return undefined;
-  const relation = RELATION_ALIASES[groups.relation.toLowerCase()];
-  const seed = groups.seed.trim().replace(/\s+/g, " ");
-  if (!relation || !seed || !LINK_RELATIONS.includes(relation))
-    return undefined;
-  return { relation, seed, direction: "incoming" };
-}
-
-function searchEmbeddingTimeoutMs(): number {
+setSearchEmbeddingTimeoutReader(() => {
   const raw =
     process.env.OPENBRAIN_SEARCH_EMBEDDING_TIMEOUT_MS ??
     process.env.SEARCH_EMBEDDING_TIMEOUT_MS;
-  if (!raw) return DEFAULT_SEARCH_EMBEDDING_TIMEOUT_MS;
-  const parsed = parseInt(raw, 10);
-  return Number.isNaN(parsed) || parsed < 1
-    ? DEFAULT_SEARCH_EMBEDDING_TIMEOUT_MS
-    : parsed;
-}
+  if (!raw) return undefined;
+  return parseInt(raw, 10);
+});
 
-async function generateSearchEmbedding(
-  deps: ToolDeps,
-  query: string,
-): Promise<number[] | null> {
-  const timeoutMs = searchEmbeddingTimeoutMs();
-  const controller = new AbortController();
-  let timeoutId: ReturnType<typeof setTimeout> | undefined;
+// Names whose legacy call form the server/ module still has verbatim.
+export {
+  ALL_TABLES,
+  mergeFallbackSearchRows,
+  readableSearchTables,
+  registerSearchBrain,
+  rrfMerge,
+  TIER_BOOST,
+  type ExplicitLink,
+  type SearchMode,
+  type SearchRow,
+  type SearchTable,
+  type SourceRef,
+  type SourceScope,
+} from "../../server/tools/legacy-search/search-brain.ts";
 
-  return traceRetrievalSpan({
-    name: "retrieval.embedding",
-    input: { query, timeout_ms: timeoutMs },
-    metadata: { stage: "candidate_generation" },
-    run: async () => {
-      try {
-        return await Promise.race([
-          deps.embedFn(query, undefined, { signal: controller.signal }),
-          new Promise<null>((resolve) => {
-            timeoutId = setTimeout(() => {
-              controller.abort();
-              logger.warn("search_embedding_timeout", {
-                timeoutMs,
-                queryLength: query.length,
-              });
-              resolve(null);
-            }, timeoutMs);
-          }),
-        ]);
-      } finally {
-        if (timeoutId) clearTimeout(timeoutId);
-      }
-    },
-    output: (embedding) => ({
-      generated: embedding !== null,
-      dimensions: embedding?.length ?? 0,
-    }),
-  });
-}
+// Names the server/ module now takes as one options object; the positional
+// wrappers below restore the legacy form.
+import {
+  executeSearch as executeSearchOptions,
+  executeSearchWithScopedSharedFallback as executeSearchWithScopedSharedFallbackOptions,
+  executeSearchWithSharedFallback as executeSearchWithSharedFallbackOptions,
+  trackUsage as trackUsageOptions,
+  type ExecuteSearchOptions,
+  type ExecuteSearchTuning,
+  type SearchMode,
+  type SearchRow,
+  type SearchTable,
+  type SourceScope,
+  type TrackUsageOptions,
+} from "../../server/tools/legacy-search/search-brain.ts";
 
-/** Gentle recency factor: today=1.0, 30d=0.97, 90d=0.92, 365d=0.73 */
-function recencyFactor(createdAt: string): number {
-  const ms = new Date(createdAt).getTime();
-  if (isNaN(ms)) return 1.0;
-  const ageDays = (Date.now() - ms) / (1000 * 60 * 60 * 24);
-  return 1 / (1 + Math.max(0, ageDays) * 0.001);
-}
+// The tier type is taken from the moved options type rather than imported from
+// `../types.ts`: an L5 adapter's relative dependencies all resolve under
+// server/ (rule M9), and this is the same `Tier` either way.
+type Tier = NonNullable<ExecuteSearchOptions["tier"]>;
 
-export interface ExplicitLink {
-  id: string;
-  direction: "outgoing" | "incoming";
-  relation: LinkRelation;
-  weight: number;
-  linked_type: string;
-  linked_id: string;
-  linked_name?: string | null;
-  canonical_id?: string | null;
-  metadata: Record<string, unknown>;
-  created_at: string;
-}
+// --- legacy positional call forms (rule M9/M15) -------------------------
+//
+// server/ takes one options object per function; every legacy src/ caller
+// still passes the historical positional list. These wrappers are that
+// translation and nothing else. The tuple element labels are the former
+// parameter names, in the former order, so the two forms read identically.
 
-export interface SourceRef {
-  source: "brain";
-  type: string;
-  id: string;
-  namespace?: string;
-  created_by?: string | null;
-  created_at?: string;
-  last_updated_at?: string;
-  label: string;
-  preview: string;
-}
-
-export interface SearchRow {
-  source_type: string;
-  id: string;
-  namespace?: string;
-  content_preview: string | null;
-  tags: string[] | null;
-  created_by?: string | null;
-  created_at: string;
-  updated_at?: string | null;
-  usefulness: number;
-  tier?: string;
-  distance?: number;
-  fts_rank?: number;
-  access_count?: number;
-  explicit_links?: ExplicitLink[];
-  source_ref?: SourceRef;
-  extracted_metadata?: {
-    topics?: string[];
-    people?: string[];
-    action_items?: string[];
-    dates?: string[];
-    // Deterministic, content-free structural keys the write-time extractor
-    // emits alongside the semantic fields. The digest, algorithm tag, and byte
-    // length reveal no source excerpt.
-    content_hash?: string;
-    hash_version?: string;
-    byte_length?: number;
-  };
-}
-
-function rowEvidence(row: SearchRow): Record<string, unknown> {
-  return {
-    row_id: row.id,
-    source_type: row.source_type,
-    namespace: row.namespace ?? null,
-    content_preview: row.content_preview ?? null,
-    distance: row.distance ?? null,
-    similarity: row.distance === undefined ? null : 1 - row.distance,
-    bm25_score: row.fts_rank ?? null,
-    usefulness: row.usefulness,
-    tier: row.tier ?? null,
-  };
-}
-
-function rowsEvidence(rows: readonly SearchRow[]): Record<string, unknown> {
-  return {
-    count: rows.length,
-    row_ids: rows.map((row) => row.id),
-    candidates: rows.map(rowEvidence),
-  };
-}
-
-function rowIdsEvidence(rows: readonly SearchRow[]): Record<string, unknown> {
-  return { count: rows.length, row_ids: rows.map((row) => row.id) };
-}
-
-const HAS_EXTRACTED_METADATA: Set<Table> = new Set(["thoughts", "decisions"]);
-
-type LinkRow = {
-  id: string;
-  from_type: string;
-  from_id: string;
-  to_type: string;
-  to_id: string;
-  relation: LinkRelation;
-  weight: number;
-  metadata: Record<string, unknown>;
-  created_at: string;
-  from_name: string | null;
-  from_canonical_id: string | null;
-  to_name: string | null;
-  to_canonical_id: string | null;
-};
-
-function linkKey(type: string, id: string): string {
-  return `${type}:${id}`;
-}
-
-function toIsoString(value: unknown): string | undefined {
-  if (typeof value !== "string" && !(value instanceof Date)) return undefined;
-  const date = new Date(value);
-  if (Number.isNaN(date.getTime())) return undefined;
-  return date.toISOString();
-}
-
-function withSourceRefs(rows: SearchRow[]): SearchRow[] {
-  return rows.map((row) => ({
-    ...row,
-    source_ref: {
-      source: "brain",
-      type: row.source_type,
-      id: row.id,
-      namespace: row.namespace,
-      created_by: row.created_by,
-      created_at: toIsoString(row.created_at),
-      last_updated_at:
-        toIsoString(row.updated_at) ?? toIsoString(row.created_at),
-      label: (row.content_preview ?? "").slice(0, 120),
-      preview: (row.content_preview ?? "").slice(0, 300),
-    },
-  }));
-}
-
-function withCanonicalNamespaces(rows: SearchRow[]): SearchRow[] {
-  return rows.map((row) => {
-    const namespace = row.namespace
-      ? canonicalNamespace(row.namespace)
-      : row.namespace;
-    return {
-      ...row,
-      namespace,
-      source_ref: row.source_ref
-        ? {
-            ...row.source_ref,
-            namespace: row.source_ref.namespace
-              ? canonicalNamespace(row.source_ref.namespace)
-              : row.source_ref.namespace,
-          }
-        : row.source_ref,
-    };
-  });
-}
-
-function dedupeSearchRows(rows: SearchRow[]): SearchRow[] {
-  const seen = new Set<string>();
-  const deduped: SearchRow[] = [];
-  for (const row of rows) {
-    const key = `${row.source_type}:${row.id}`;
-    if (seen.has(key)) continue;
-    seen.add(key);
-    deduped.push(row);
-  }
-  return deduped;
-}
-
-function fallbackDedupeKey(row: SearchRow): string {
-  const preview = row.content_preview?.replace(/\s+/g, " ").trim();
-  if (preview) {
-    return `${row.source_type}:content:${preview}`;
-  }
-  return `${row.source_type}:id:${row.id}`;
-}
-
-function dedupeFallbackSearchRows(rows: SearchRow[]): SearchRow[] {
-  const seen = new Set<string>();
-  const deduped: SearchRow[] = [];
-  for (const row of rows) {
-    const key = fallbackDedupeKey(row);
-    if (seen.has(key)) continue;
-    seen.add(key);
-    deduped.push(row);
-  }
-  return deduped;
-}
-
-type FallbackClassification = {
-  row: SearchRow;
-  chosen: boolean;
-  filtered_by: "fallback_duplicate" | "fallback_limit" | null;
-};
-
-function selectedFallbackRows(
-  primary: SearchRow[],
-  legacy: SearchRow[],
-  limit: number,
-): SearchRow[] {
-  if (legacy.length === 0) return primary.slice(0, limit);
-  const firstLegacy = legacy[0];
-  if (!firstLegacy) return primary.slice(0, limit);
-  if (primary.length >= limit) {
-    return [...primary.slice(0, Math.max(0, limit - 1)), firstLegacy];
-  }
-  return [...primary, ...legacy.slice(0, limit - primary.length)];
-}
-
-function fallbackFilteredBy(
-  chosen: boolean,
-  duplicate: boolean,
-): FallbackClassification["filtered_by"] {
-  if (chosen) return null;
-  return duplicate ? "fallback_duplicate" : "fallback_limit";
-}
-
-function computeFallbackSearchRows(
-  primaryRows: SearchRow[],
-  legacyRows: SearchRow[],
-  limit: number,
-): { rows: SearchRow[]; classifications: FallbackClassification[] } {
-  const primary = dedupeFallbackSearchRows(primaryRows);
-  const primaryKeys = new Set(primary.map(fallbackDedupeKey));
-  const legacy = dedupeFallbackSearchRows(
-    legacyRows.filter((row) => !primaryKeys.has(fallbackDedupeKey(row))),
-  );
-  const rows = selectedFallbackRows(primary, legacy, limit);
-  const selectedKeys = new Set(rows.map(fallbackDedupeKey));
-  const seen = new Set<string>();
-  const classifications = [...primaryRows, ...legacyRows].map((row, index) => {
-    const key = fallbackDedupeKey(row);
-    const duplicate =
-      seen.has(key) ||
-      (index >= primaryRows.length && primaryKeys.has(key));
-    seen.add(key);
-    const chosen = !duplicate && selectedKeys.has(key);
-    return {
-      row,
-      chosen,
-      filtered_by: fallbackFilteredBy(chosen, duplicate),
-    } satisfies FallbackClassification;
-  });
-  return { rows, classifications };
-}
-
-export function mergeFallbackSearchRows(
-  primaryRows: SearchRow[],
-  legacyRows: SearchRow[],
-  limit: number,
-): SearchRow[] {
-  const result = traceRetrievalSpanSync({
-    name: "retrieval.fallback_dedupe",
-    input: {
-      limit,
-      primary: rowsEvidence(primaryRows),
-      legacy: rowsEvidence(legacyRows),
-    },
-    metadata: {
-      stage: "filtering",
-      filter_names: ["fallback_duplicate", "fallback_limit"],
-    },
-    run: () => computeFallbackSearchRows(primaryRows, legacyRows, limit),
-    output: ({ rows, classifications }) => ({
-      candidate_count: classifications.length,
-      selected_count: rows.length,
-      selected_row_ids: rows.map((row) => row.id),
-      candidates: classifications.map(({ row, chosen, filtered_by }) => ({
-        ...rowEvidence(row),
-        chosen,
-        filtered_by,
-      })),
-    }),
-  });
-  return result.rows;
-}
-
-function appendNamespaceParam(
-  params: unknown[],
-  namespace?: NamespaceFilter,
-): number | undefined {
-  if (namespace === undefined) return undefined;
-  params.push(namespace);
-  return params.length;
-}
-
-function paramRef(index: number): string {
-  if (!Number.isInteger(index) || index < 1) {
-    throw new Error(`Invalid SQL parameter index: ${index}`);
-  }
-  return `$${index}`;
-}
-
-async function attachExplicitLinks(
-  deps: ToolDeps,
-  rows: SearchRow[],
-  namespace?: NamespaceFilter,
-): Promise<SearchRow[]> {
-  if (rows.length === 0) return rows;
-
-  const resultTypes = rows.map((row) => row.source_type);
-  const resultIds = rows.map((row) => row.id);
-  const params: unknown[] = [resultTypes, resultIds];
-  const namespaceParamIndex = appendNamespaceParam(params, namespace);
-  const namespaceFilter = namespaceParamIndex
-    ? Array.isArray(namespace)
-      ? ` AND namespace = ANY(${paramRef(namespaceParamIndex)}::text[])`
-      : ` AND namespace = ${paramRef(namespaceParamIndex)}`
-    : "";
-
-  try {
-    const { rows: linkRows } = await deps.pool.query<LinkRow>(
-      `SELECT
-         l.id, l.from_type, l.from_id, l.to_type, l.to_id, l.relation, l.weight, l.metadata, l.created_at,
-         from_entity.name AS from_name,
-         from_entity.canonical_id AS from_canonical_id,
-         to_entity.name AS to_name,
-         to_entity.canonical_id AS to_canonical_id
-       FROM ob_links l
-       LEFT JOIN ob_entities from_entity
-         ON l.from_type = 'entity'
-        AND from_entity.id = l.from_id
-        AND from_entity.namespace = l.namespace
-        AND from_entity.archived_at IS NULL
-       LEFT JOIN ob_entities to_entity
-         ON l.to_type = 'entity'
-        AND to_entity.id = l.to_id
-        AND to_entity.namespace = l.namespace
-        AND to_entity.archived_at IS NULL
-       WHERE (
-         (l.from_type, l.from_id) IN (
-           SELECT result_type, result_id
-           FROM unnest($1::text[], $2::uuid[]) AS result_refs(result_type, result_id)
-         )
-         OR (l.to_type, l.to_id) IN (
-           SELECT result_type, result_id
-           FROM unnest($1::text[], $2::uuid[]) AS result_refs(result_type, result_id)
-         )
-       )
-         AND l.archived_at IS NULL
-         AND (l.from_type <> 'entity' OR from_entity.id IS NOT NULL)
-         AND (l.to_type <> 'entity' OR to_entity.id IS NOT NULL)${namespaceFilter.replaceAll("namespace", "l.namespace")}
-       ORDER BY l.weight DESC, l.created_at DESC
-       LIMIT 200`,
-      params,
-    );
-
-    if (linkRows.length === 200) {
-      logger.warn("explicit_links_truncated", {
-        result_count: rows.length,
-        link_limit: 200,
-      });
-    }
-
-    const linksByResult = new Map<string, ExplicitLink[]>();
-    for (const link of linkRows) {
-      const outgoingKey = linkKey(link.from_type, link.from_id);
-      const incomingKey = linkKey(link.to_type, link.to_id);
-      const outgoing = linksByResult.get(outgoingKey) ?? [];
-      outgoing.push({
-        id: link.id,
-        direction: "outgoing",
-        relation: link.relation,
-        weight: link.weight,
-        linked_type: link.to_type,
-        linked_id: link.to_id,
-        linked_name: link.to_name,
-        canonical_id: link.to_canonical_id,
-        metadata: link.metadata ?? {},
-        created_at: new Date(link.created_at).toISOString(),
-      });
-      linksByResult.set(outgoingKey, outgoing);
-
-      const incoming = linksByResult.get(incomingKey) ?? [];
-      incoming.push({
-        id: link.id,
-        direction: "incoming",
-        relation: link.relation,
-        weight: link.weight,
-        linked_type: link.from_type,
-        linked_id: link.from_id,
-        linked_name: link.from_name,
-        canonical_id: link.from_canonical_id,
-        metadata: link.metadata ?? {},
-        created_at: new Date(link.created_at).toISOString(),
-      });
-      linksByResult.set(incomingKey, incoming);
-    }
-
-    return rows.map((row) => ({
-      ...row,
-      explicit_links: linksByResult.get(linkKey(row.source_type, row.id)) ?? [],
-    }));
-  } catch (err) {
-    logger.warn("explicit_links_lookup_failed", {
-      error: err instanceof Error ? err.message : String(err),
-    });
-    return rows.map((row) => ({ ...row, explicit_links: [] }));
-  }
-}
-
-function buildTableCTE(
-  table: SearchTable,
-  perTableLimit: number,
-  tier?: Tier,
-  namespaceParamIndex?: number,
-  namespaceIsArray = false,
-  sourceScopeParamIndex?: number,
-): string {
-  if (tier && !VALID_TIERS.has(tier)) throw new Error(`Invalid tier: ${tier}`);
-  if (table === "entities") {
-    const tierFilter = tier && tier !== "warm" ? " AND FALSE" : "";
-    const nsFilter = namespaceParamIndex
-      ? namespaceIsArray
-        ? ` AND e.namespace = ANY(${paramRef(namespaceParamIndex)}::text[])`
-        : ` AND e.namespace = ${paramRef(namespaceParamIndex)}`
-      : "";
-    return `entities_results AS (
-  SELECT
-    'entity' AS source_type,
-    e.id,
-    e.namespace,
-    e.entity_type || ': ' || e.name ||
-      CASE WHEN e.canonical_id IS NOT NULL THEN ' (' || e.canonical_id || ')' ELSE '' END AS content_preview,
-    NULL::text[] AS tags,
-    e.created_by,
-    e.created_at,
-    e.updated_at,
-    'warm'::text AS tier,
-    e.embedding <=> (SELECT emb FROM query_embedding) AS distance,
-    0.5 AS usefulness,
-    0 AS access_count,
-    NULL::jsonb AS extracted_metadata
-  FROM ob_entities e
-  WHERE e.embedding IS NOT NULL AND e.archived_at IS NULL${tierFilter}${nsFilter}
-  ORDER BY e.embedding <=> (SELECT emb FROM query_embedding) ASC
-  LIMIT ${perTableLimit}
-)`;
-  }
-  if (table === "session_events") {
-    // A source scope cannot be evaluated against a corpus with no source_refs;
-    // the caller drops this table from the list in that case (see
-    // executeSearchInternal), rather than ignoring the scope and leaking
-    // out-of-scope rows into a scoped query.
-    const tierFilter = tier ? ` AND se.importance = '${tier}'` : "";
-    const nsFilter = namespaceParamIndex
-      ? namespaceIsArray
-        ? ` AND sl.namespace = ANY(${paramRef(namespaceParamIndex)}::text[])`
-        : ` AND sl.namespace = ${paramRef(namespaceParamIndex)}`
-      : "";
-    return `session_events_results AS (
-  SELECT
-    '${SESSION_EVENTS_SOURCE_LABEL}' AS source_type,
-    se.id,
-    sl.namespace,
-    se.content AS content_preview,
-    NULL::text[] AS tags,
-    se.created_by,
-    se.created_at,
-    se.created_at AS updated_at,
-    se.importance AS tier,
-    se.embedding <=> (SELECT emb FROM query_embedding) AS distance,
-    0.5 AS usefulness,
-    0 AS access_count,
-    se.metadata AS extracted_metadata
-  FROM ob_session_events se
-  JOIN ob_session_lanes sl ON sl.id = se.lane_id
-  WHERE se.embedding IS NOT NULL${tierFilter}${nsFilter}
-  ORDER BY se.embedding <=> (SELECT emb FROM query_embedding) ASC
-  LIMIT ${perTableLimit}
-)`;
-  }
-  const alias = TABLE_ALIAS[table];
-  const label = SOURCE_LABELS[table];
-  const preview = CONTENT_PREVIEW[table];
-  const cteName = `${table}_results`;
-  const tierFilter = tier ? ` AND ${alias}.tier = '${tier}'` : "";
-  const nsFilter = namespaceParamIndex
-    ? namespaceIsArray
-      ? ` AND ${alias}.namespace = ANY(${paramRef(namespaceParamIndex)}::text[])`
-      : ` AND ${alias}.namespace = ${paramRef(namespaceParamIndex)}`
-    : "";
-  const sourceScopeFilter = sourceScopeFilterSql(alias, sourceScopeParamIndex);
-  const metaCol = HAS_EXTRACTED_METADATA.has(table)
-    ? `${alias}.extracted_metadata`
-    : "NULL::jsonb AS extracted_metadata";
-
-  return `${cteName} AS (
-  SELECT
-    '${label}' AS source_type,
-    ${alias}.id,
-    ${alias}.namespace,
-    ${preview} AS content_preview,
-    ${alias}.tags,
-    ${alias}.created_by,
-    ${alias}.created_at,
-    ${alias}.updated_at,
-    ${alias}.tier,
-    ${alias}.embedding <=> (SELECT emb FROM query_embedding) AS distance,
-    COALESCE(${alias}.usefulness_score, 0.5) AS usefulness,
-    COALESCE(${alias}.access_count, 0) AS access_count,
-    ${metaCol}
-  FROM ${table} ${alias}
-  WHERE ${alias}.embedding IS NOT NULL AND ${alias}.archived_at IS NULL${tierFilter}${nsFilter}${sourceScopeFilter}
-  ORDER BY ${alias}.embedding <=> (SELECT emb FROM query_embedding) ASC
-  LIMIT ${perTableLimit}
-)`;
-}
-
-/**
- * Build the lexical (FTS) match + rank expressions for one table under a chosen
- * text-search configuration.
- *
- * english (default): use the GIN-indexed stored `search_vector` column exactly
- * as before -- byte-identical to the pre-#341 behavior and index-fast.
- *
- * non-english supported config: recompute `to_tsvector(<config>, <source text>)`
- * on the fly against the same columns the stored column indexes, so the query
- * arm and the analyzed text share one configuration (correct stemming, no
- * index/query mismatch, no migration). `config` is an allowlist-validated
- * FtsConfig; ftsConfigLiteral re-asserts that before it is interpolated.
- */
-function ftsMatchExpressions(
-  table: Table,
-  config: FtsConfig,
-): {
-  vectorSql: string;
-  querySql: string;
-} {
-  const querySql = `plainto_tsquery('${ftsConfigLiteral(config)}', (SELECT q FROM fts_query))`;
-  if (config === "english") {
-    const alias = TABLE_ALIAS[table];
-    return { vectorSql: `${alias}.search_vector`, querySql };
-  }
-  const vectorSql = `to_tsvector('${ftsConfigLiteral(config)}', ${FTS_SOURCE_TEXT[table]})`;
-  return { vectorSql, querySql };
-}
-
-function buildFtsCTE(
-  table: SearchTable,
-  perTableLimit: number,
-  ftsConfig: FtsConfig,
-  tier?: Tier,
-  namespaceParamIndex?: number,
-  namespaceIsArray = false,
-  sourceScopeParamIndex?: number,
-): string {
-  if (tier && !VALID_TIERS.has(tier)) throw new Error(`Invalid tier: ${tier}`);
-  if (table === "entities") {
-    const tierFilter = tier && tier !== "warm" ? " AND FALSE" : "";
-    const nsFilter = namespaceParamIndex
-      ? namespaceIsArray
-        ? ` AND e.namespace = ANY(${paramRef(namespaceParamIndex)}::text[])`
-        : ` AND e.namespace = ${paramRef(namespaceParamIndex)}`
-      : "";
-    return `entities_fts AS (
-  SELECT
-    'entity' AS source_type,
-    e.id,
-    e.namespace,
-    e.entity_type || ': ' || e.name ||
-      CASE WHEN e.canonical_id IS NOT NULL THEN ' (' || e.canonical_id || ')' ELSE '' END AS content_preview,
-    NULL::text[] AS tags,
-    e.created_by,
-    e.created_at,
-    e.updated_at,
-    'warm'::text AS tier,
-    1.0 AS fts_rank,
-    0.5 AS usefulness,
-    0 AS access_count,
-    NULL::jsonb AS extracted_metadata
-  FROM ob_entities e
-  WHERE (
-      e.name ILIKE '%' || (SELECT q FROM fts_query) || '%'
-      OR e.entity_type ILIKE '%' || (SELECT q FROM fts_query) || '%'
-      OR e.canonical_id ILIKE '%' || (SELECT q FROM fts_query) || '%'
-      OR e.metadata::text ILIKE '%' || (SELECT q FROM fts_query) || '%'
-    )
-    AND e.archived_at IS NULL${tierFilter}${nsFilter}
-  ORDER BY e.updated_at DESC
-  LIMIT ${perTableLimit}
-)`;
-  }
-  if (table === "session_events") {
-    // ob_session_events has no stored `search_vector` generated column, so
-    // there is no tsvector to match against and no GIN index to ride. The
-    // lexical arm matches `content` with ILIKE, exactly as the entities arm
-    // above does for its own unindexed text. This is substring matching, not
-    // stemming: it is the honest capability of a column that does not exist,
-    // and it is stated here so a reader does not assume FTS parity. The
-    // `ftsConfig` argument is deliberately unused for this table -- a text
-    // search configuration has nothing to configure without a tsvector.
-    const tierFilter = tier ? ` AND se.importance = '${tier}'` : "";
-    const nsFilter = namespaceParamIndex
-      ? namespaceIsArray
-        ? ` AND sl.namespace = ANY(${paramRef(namespaceParamIndex)}::text[])`
-        : ` AND sl.namespace = ${paramRef(namespaceParamIndex)}`
-      : "";
-    return `session_events_fts AS (
-  SELECT
-    '${SESSION_EVENTS_SOURCE_LABEL}' AS source_type,
-    se.id,
-    sl.namespace,
-    se.content AS content_preview,
-    NULL::text[] AS tags,
-    se.created_by,
-    se.created_at,
-    se.created_at AS updated_at,
-    se.importance AS tier,
-    1.0 AS fts_rank,
-    0.5 AS usefulness,
-    0 AS access_count,
-    se.metadata AS extracted_metadata
-  FROM ob_session_events se
-  JOIN ob_session_lanes sl ON sl.id = se.lane_id
-  WHERE se.content ILIKE '%' || (SELECT q FROM fts_query) || '%'${tierFilter}${nsFilter}
-  ORDER BY se.created_at DESC
-  LIMIT ${perTableLimit}
-)`;
-  }
-  const alias = TABLE_ALIAS[table];
-  const label = SOURCE_LABELS[table];
-  const preview = CONTENT_PREVIEW[table];
-  const cteName = `${table}_fts`;
-  const tierFilter = tier ? ` AND ${alias}.tier = '${tier}'` : "";
-  const nsFilter = namespaceParamIndex
-    ? namespaceIsArray
-      ? ` AND ${alias}.namespace = ANY(${paramRef(namespaceParamIndex)}::text[])`
-      : ` AND ${alias}.namespace = ${paramRef(namespaceParamIndex)}`
-    : "";
-  const sourceScopeFilter = sourceScopeFilterSql(alias, sourceScopeParamIndex);
-
-  const metaCol = HAS_EXTRACTED_METADATA.has(table)
-    ? `${alias}.extracted_metadata`
-    : "NULL::jsonb AS extracted_metadata";
-
-  const { vectorSql, querySql } = ftsMatchExpressions(table, ftsConfig);
-
-  return `${cteName} AS (
-  SELECT
-    '${label}' AS source_type,
-    ${alias}.id,
-    ${alias}.namespace,
-    ${preview} AS content_preview,
-    ${alias}.tags,
-    ${alias}.created_by,
-    ${alias}.created_at,
-    ${alias}.updated_at,
-    ${alias}.tier,
-    ts_rank_cd(${vectorSql}, ${querySql}) AS fts_rank,
-    COALESCE(${alias}.usefulness_score, 0.5) AS usefulness,
-    COALESCE(${alias}.access_count, 0) AS access_count,
-    ${metaCol}
-  FROM ${table} ${alias}
-  WHERE ${vectorSql} @@ ${querySql}
-    AND ${alias}.archived_at IS NULL${tierFilter}${nsFilter}${sourceScopeFilter}
-  ORDER BY fts_rank DESC
-  LIMIT ${perTableLimit}
-)`;
-}
-
-function buildRelationalHydrationSelect(
-  table: Table,
-  direction: RelationalDirection,
-  tier?: Tier,
-): string {
-  if (tier && !VALID_TIERS.has(tier)) throw new Error(`Invalid tier: ${tier}`);
-  const alias = TABLE_ALIAS[table];
-  const label = SOURCE_LABELS[table];
-  const preview = CONTENT_PREVIEW[table];
-  const tierFilter = tier ? ` AND ${alias}.tier = '${tier}'` : "";
-  const metaCol = HAS_EXTRACTED_METADATA.has(table)
-    ? `${alias}.extracted_metadata`
-    : "NULL::jsonb AS extracted_metadata";
-  const linkJoin =
-    direction === "incoming"
-      ? `l.to_type = 'entity'
-   AND l.to_id = seed.id`
-      : `l.from_type = 'entity'
-   AND l.from_id = seed.id`;
-  const targetJoin =
-    direction === "incoming"
-      ? `l.from_type = '${label}'
-   AND ${alias}.id = l.from_id`
-      : `l.to_type = '${label}'
-   AND ${alias}.id = l.to_id`;
-
-  return `SELECT
-    '${label}' AS source_type,
-    ${alias}.id,
-    ${alias}.namespace,
-    ${preview} AS content_preview,
-    ${alias}.tags,
-    ${alias}.created_by,
-    ${alias}.created_at,
-    ${alias}.updated_at,
-    ${alias}.tier,
-    NULL::double precision AS distance,
-    GREATEST(l.weight, 0)::double precision AS fts_rank,
-    COALESCE(${alias}.usefulness_score, 0.5) AS usefulness,
-    COALESCE(${alias}.access_count, 0) AS access_count,
-    ${metaCol}
-  FROM relational_graph_seed seed
-  JOIN ob_links l
-    ON ${linkJoin}
-   AND l.namespace = seed.namespace
-   AND l.relation = $2
-   AND l.archived_at IS NULL
-  JOIN ${table} ${alias}
-    ON ${targetJoin}
-   AND ${alias}.namespace = l.namespace
-   AND ${alias}.archived_at IS NULL${tierFilter}`;
-}
-
-async function relationalGraphSearch(
-  deps: ToolDeps,
+type ExecuteArgs = [
+  deps: Parameters<typeof executeSearchOptions>[0]["deps"],
   accessibleTables: SearchTable[],
   query: string,
-  fetchLimit: number,
+  count: number,
+  mode?: SearchMode,
   tier?: Tier,
-  namespace?: NamespaceFilter,
-): Promise<SearchRow[]> {
-  const parsed = parseRelationalQuery(query);
-  if (!parsed) return [];
-  // Relational hydration walks `ob_links` from a seed entity into the physical
-  // content tables. `entities` is the seed side, and `session_events` has no
-  // rows in the link graph, so neither is a hydration target.
-  const targetTables = accessibleTables.filter(
-    (table): table is Table =>
-      table !== "entities" && table !== "session_events",
-  );
-  if (targetTables.length === 0) return [];
-
-  const params: unknown[] = [parsed.seed, parsed.relation, fetchLimit];
-  const namespaceParamIndex = appendNamespaceParam(params, namespace);
-  const namespaceIsArray = Array.isArray(namespace);
-  const namespaceFilter = namespaceParamIndex
-    ? namespaceIsArray
-      ? ` AND e.namespace = ANY(${paramRef(namespaceParamIndex)}::text[])`
-      : ` AND e.namespace = ${paramRef(namespaceParamIndex)}`
-    : "";
-  const hydrationSql = targetTables
-    .map((table) =>
-      buildRelationalHydrationSelect(table, parsed.direction, tier),
-    )
-    .join("\nUNION ALL\n");
-
-  try {
-    const rows = await traceRetrievalSpan({
-      name: "retrieval.graph_query",
-      input: {
-        relation: parsed.relation,
-        direction: parsed.direction,
-        seed: parsed.seed,
-        fetch_limit: fetchLimit,
-        tier,
-        namespace,
-      },
-      metadata: {
-        stage: "candidate_generation",
-        filter_names: ["namespace", "relation", "archived_at"],
-      },
-      run: async () => {
-        const result = await deps.pool.query<SearchRow>(
-          `WITH relational_graph_seed AS (
-         SELECT e.id, e.namespace
-         FROM ob_entities e
-         WHERE (
-             lower(e.name) = lower($1)
-             OR lower(COALESCE(e.canonical_id, '')) = lower($1)
-           )
-           AND e.archived_at IS NULL${namespaceFilter}
-         ORDER BY e.updated_at DESC
-         LIMIT 5
-       )
-       SELECT *
-       FROM (
-${hydrationSql}
-       ) relational_graph_rows
-       ORDER BY fts_rank DESC, created_at DESC
-       LIMIT $3`,
-          params,
-        );
-        return withSourceRefs(result.rows);
-      },
-      output: rowsEvidence,
-    });
-    logger.info("search_relational_graph", {
-      relation: parsed.relation,
-      direction: parsed.direction,
-      seed_length: parsed.seed.length,
-      target_tables: targetTables,
-      candidate_count: rows.length,
-    });
-    return rows;
-  } catch (err) {
-    logger.warn("search_relational_graph_failed", {
-      relation: parsed.relation,
-      direction: parsed.direction,
-      seed_length: parsed.seed.length,
-      error: err instanceof Error ? err.message : String(err),
-    });
-    return [];
-  }
-}
-
-async function vectorSearch(
-  deps: ToolDeps,
-  accessibleTables: SearchTable[],
-  embedding: number[],
-  fetchLimit: number,
-  tier?: Tier,
-  offset = 0,
-  namespace?: NamespaceFilter,
+  offset?: number,
+  namespace?: string | string[],
+  includeLinks?: boolean,
   sourceScope?: SourceScope,
+  tuning?: ExecuteSearchTuning,
+];
+
+function toExecuteOptions(args: ExecuteArgs): ExecuteSearchOptions {
+  const [
+    deps,
+    accessibleTables,
+    query,
+    count,
+    mode,
+    tier,
+    offset,
+    namespace,
+    includeLinks,
+    sourceScope,
+    tuning,
+  ] = args;
+  return {
+    deps,
+    accessibleTables,
+    query,
+    limit: count,
+    mode,
+    tier,
+    offset,
+    namespace,
+    includeLinks,
+    sourceScope,
+    tuning,
+  };
+}
+
+export function executeSearch(...args: ExecuteArgs): Promise<SearchRow[]> {
+  return executeSearchOptions(toExecuteOptions(args));
+}
+
+export function executeSearchWithSharedFallback(
+  ...args: ExecuteArgs
 ): Promise<SearchRow[]> {
-  const perTableLimit = fetchLimit;
-  const params = [toSql(embedding), fetchLimit, offset];
-  const namespaceParamIndex = appendNamespaceParam(params, namespace);
-  const sourceScopeParamIndex = appendSourceScopeParam(params, sourceScope);
-  const namespaceIsArray = Array.isArray(namespace);
-  const ctes = accessibleTables.map((t) =>
-    buildTableCTE(
-      t,
-      perTableLimit,
-      tier,
-      namespaceParamIndex,
-      namespaceIsArray,
-      sourceScopeParamIndex,
-    ),
-  );
-  const cteNames = accessibleTables.map((t) => `${t}_results`);
-  const unionAll = cteNames
-    .map((name) => `SELECT * FROM ${name}`)
-    .join("\nUNION ALL\n");
-
-  const sql = `WITH query_embedding AS (
-  SELECT $1::halfvec(768) AS emb
-),
-${ctes.join(",\n")}
-SELECT * FROM (
-${unionAll}
-) AS combined
-ORDER BY (distance * ${VECTOR_WEIGHT} + (1.0 - COALESCE(usefulness, 0.5)) * ${USEFULNESS_WEIGHT} + EXTRACT(EPOCH FROM (NOW() - created_at)) / 86400.0 * ${AGE_WEIGHT}) ASC
-LIMIT $2 OFFSET $3`;
-
-  return traceRetrievalSpan({
-    name: "retrieval.vector_query",
-    input: {
-      accessible_tables: accessibleTables,
-      fetch_limit: fetchLimit,
-      offset,
-      tier,
-      namespace,
-      source_scope: sourceScope,
-    },
-    metadata: {
-      stage: "candidate_generation",
-      filter_names: ["namespace", "tier", "source_scope", "archived_at"],
-    },
-    run: async () => {
-      const { rows } = await deps.pool.query(sql, params);
-      return withSourceRefs(rows as SearchRow[]);
-    },
-    output: rowsEvidence,
-  });
+  return executeSearchWithSharedFallbackOptions(toExecuteOptions(args));
 }
 
-/**
- * Execute one FTS statement under the cost policy for its configuration
- * (the query execution boundary for the lexical arm).
- *
- * english (default): a plain pooled query. The GIN-indexed stored-column path
- * needs no extra bound and stays byte-identical to the pre-#341 execution.
- *
- * non-default config: the match arm recomputes `to_tsvector` per row with no
- * index (see ftsMatchExpressions), so the statement runs inside a transaction
- * bounded by a transaction-scoped `SET LOCAL statement_timeout`. The bound
- * comes from OPENBRAIN_FTS_STATEMENT_TIMEOUT_MS validated to a positive
- * integer (default 5000 ms) by ftsStatementTimeoutMs -- the raw env string is
- * never interpolated. `SET` accepts no bind parameters, so the vetted number
- * (and only the number) is inlined, mirroring the allowlist-then-interpolate
- * discipline used for table names and the regconfig literal. SET LOCAL scopes
- * the timeout to this transaction, so the pooled connection is returned clean.
- */
-async function runBoundedFtsQuery(
-  deps: ToolDeps,
-  sql: string,
-  params: unknown[],
-  ftsConfig: FtsConfig,
-): Promise<{ rows: unknown[] }> {
-  if (ftsConfig === DEFAULT_FTS_CONFIG) {
-    return deps.pool.query(sql, params as unknown[] as never[]);
-  }
-  const timeoutMs = ftsStatementTimeoutMs();
-  const client = await deps.pool.connect();
-  try {
-    await client.query("BEGIN");
-    const result = await (async () => {
-      await client.query(`SET LOCAL statement_timeout = ${timeoutMs}`);
-      const queried = await client.query(sql, params as unknown[] as never[]);
-      await client.query("COMMIT");
-      return queried;
-    })();
-    client.release();
-    return result;
-  } catch (err) {
-    try {
-      await client.query("ROLLBACK");
-      client.release();
-    } catch {
-      // The original error is the signal; a failed ROLLBACK must not mask
-      // it. Destroy the client so a connection with an aborted transaction
-      // (or a dead socket) can never be handed back to the shared pool.
-      client.release(err instanceof Error ? err : new Error(String(err)));
-    }
-    throw err;
-  }
-}
-
-async function ftsSearch(
-  deps: ToolDeps,
-  accessibleTables: SearchTable[],
-  query: string,
-  fetchLimit: number,
-  tier?: Tier,
-  offset = 0,
-  namespace?: NamespaceFilter,
-  sourceScope?: SourceScope,
-  ftsConfig: FtsConfig = DEFAULT_FTS_CONFIG,
+export function executeSearchWithScopedSharedFallback(
+  ...args: ExecuteArgs
 ): Promise<SearchRow[]> {
-  const perTableLimit = fetchLimit;
-  const params = [query, fetchLimit, offset];
-  const namespaceParamIndex = appendNamespaceParam(params, namespace);
-  const sourceScopeParamIndex = appendSourceScopeParam(params, sourceScope);
-  const namespaceIsArray = Array.isArray(namespace);
-  const ctes = accessibleTables.map((t) =>
-    buildFtsCTE(
-      t,
-      perTableLimit,
-      ftsConfig,
-      tier,
-      namespaceParamIndex,
-      namespaceIsArray,
-      sourceScopeParamIndex,
-    ),
-  );
-  const cteNames = accessibleTables.map((t) => `${t}_fts`);
-  const unionAll = cteNames
-    .map((name) => `SELECT * FROM ${name}`)
-    .join("\nUNION ALL\n");
-
-  const sql = `WITH fts_query AS (
-  SELECT $1::text AS q
-),
-${ctes.join(",\n")}
-SELECT * FROM (
-${unionAll}
-) AS combined
-ORDER BY fts_rank DESC
-LIMIT $2 OFFSET $3`;
-
-  return traceRetrievalSpan({
-    name: "retrieval.keyword_query",
-    input: {
-      query,
-      accessible_tables: accessibleTables,
-      fetch_limit: fetchLimit,
-      offset,
-      tier,
-      namespace,
-      source_scope: sourceScope,
-      fts_config: ftsConfig,
-    },
-    metadata: {
-      stage: "candidate_generation",
-      filter_names: ["namespace", "tier", "source_scope", "archived_at"],
-    },
-    run: async () => {
-      const { rows } = await runBoundedFtsQuery(deps, sql, params, ftsConfig);
-      return withSourceRefs(rows as SearchRow[]);
-    },
-    output: rowsEvidence,
-  });
+  return executeSearchWithScopedSharedFallbackOptions(toExecuteOptions(args));
 }
 
-/**
- * Reciprocal Rank Fusion: merge ranked lists from different scoring systems.
- * Items appearing in both lists get summed RRF scores (boosted).
- * Items in only one list get a single RRF score.
- * Hot entries get +0.3 boost, cold entries get -0.2, warm is unchanged.
- */
-export function rrfMerge(
-  vectorRows: SearchRow[],
-  ftsRows: SearchRow[],
-  limit: number,
-  graphRows: SearchRow[] = [],
-): SearchRow[] {
-  let ranked: Array<{ row: SearchRow; rrf: number }> = [];
-  return traceRetrievalSpanSync({
-    name: "retrieval.rank_rrf",
-    input: {
-      limit,
-      vector: rowIdsEvidence(vectorRows),
-      keyword: rowIdsEvidence(ftsRows),
-      graph: rowIdsEvidence(graphRows),
-    },
-    metadata: { stage: "scoring_ranking", filter_names: ["rrf_window"] },
-    run: () => {
-      const scoreMap = new Map<string, { row: SearchRow; rrf: number }>();
-      const accumulate = (rows: SearchRow[], multiplier = 1): void => {
-        rows.forEach((row, index) => {
-          const key = `${row.source_type}:${row.id}`;
-          const contribution = multiplier / (RRF_K + index + 1);
-          const existing = scoreMap.get(key);
-          if (existing) {
-            existing.rrf += contribution;
-            if (multiplier === 3) {
-              existing.row = { ...existing.row, explicit_links: row.explicit_links };
-            }
-          } else {
-            scoreMap.set(key, { row, rrf: contribution });
-          }
-        });
-      };
-      accumulate(vectorRows);
-      accumulate(ftsRows);
-      accumulate(graphRows, 3);
-      ranked = Array.from(scoreMap.values())
-        .map(({ row, rrf }) => ({
-          row,
-          rrf: Math.max(
-            0,
-            (rrf + TIER_BOOST[(row.tier ?? "warm") as Tier]) *
-              (TABLE_WEIGHT[row.source_type] ?? 1.0) *
-              recencyFactor(row.created_at),
-          ),
-        }))
-        .sort((a, b) => b.rrf - a.rrf);
-      return ranked.slice(0, limit).map(({ row }) => row);
-    },
-    output: (rows) => {
-      const selectedKeys = new Set(
-        rows.map((row) => `${row.source_type}:${row.id}`),
-      );
-      return {
-        candidate_count: ranked.length,
-        selected_count: rows.length,
-        selected_row_ids: rows.map((row) => row.id),
-        candidates: ranked.map(({ row, rrf }) => {
-          const chosen = selectedKeys.has(`${row.source_type}:${row.id}`);
-          return {
-            ...rowEvidence(row),
-            rank_score: rrf,
-            chosen,
-            filtered_by: chosen ? null : "rrf_window",
-          };
-        }),
-      };
-    },
-  });
-}
-
-/** Circuit breaker: stop tracking after consecutive failures to avoid log floods */
-const TRACKING_FAILURE_THRESHOLD = 5;
-let trackingConsecutiveFailures = 0;
-let trackingCircuitOpen = false;
-
-export function trackUsage(
-  deps: ToolDeps,
+type TrackUsageArgs = [
+  deps: TrackUsageOptions["deps"],
   rows: SearchRow[],
   queryText: string,
-  context = "search",
+  context?: string,
   accessedBy?: string,
-): void {
-  if (rows.length === 0) return;
-  if (trackingCircuitOpen) return;
+];
 
-  const byTable = new Map<Table, string[]>();
-  for (const row of rows) {
-    const table = LABEL_TO_TABLE[row.source_type];
-    if (!table) continue;
-    const ids = byTable.get(table) ?? [];
-    ids.push(row.id);
-    byTable.set(table, ids);
-  }
-
-  const trackingPromises: Promise<unknown>[] = [];
-  for (const [table, ids] of byTable) {
-    trackingPromises.push(
-      deps.pool.query(
-        `UPDATE ${table} SET access_count = access_count + 1, last_accessed_at = NOW() WHERE id = ANY($1)`,
-        [ids],
-      ),
-    );
-  }
-
-  // Bulk-insert into entry_access_log for all returned entries
-  const logRows = rows.filter((r) => LABEL_TO_TABLE[r.source_type]);
-  if (logRows.length > 0) {
-    const entryIds = logRows.map((r) => r.id);
-    const sourceTables = logRows.map((r) => LABEL_TO_TABLE[r.source_type]);
-    trackingPromises.push(
-      deps.pool.query(
-        `INSERT INTO entry_access_log (entry_id, source_table, accessed_at, query_text, context, accessed_by)
-           SELECT unnest($1::uuid[]), unnest($2::text[]), NOW(), $3, $4, $5`,
-        [entryIds, sourceTables, queryText, context, accessedBy ?? null],
-      ),
-    );
-  }
-
-  void Promise.allSettled(trackingPromises).then((results) => {
-    const anyFailed = results.some((r) => r.status === "rejected");
-    if (anyFailed) {
-      trackingConsecutiveFailures++;
-      if (trackingConsecutiveFailures >= TRACKING_FAILURE_THRESHOLD) {
-        trackingCircuitOpen = true;
-        logger.warn("search_tracking_circuit_open", {
-          message: `Tracking disabled after ${TRACKING_FAILURE_THRESHOLD} consecutive failures`,
-        });
-      } else {
-        const firstError = results.find((r) => r.status === "rejected") as
-          | PromiseRejectedResult
-          | undefined;
-        logger.warn("search_tracking_error", {
-          error:
-            firstError?.reason instanceof Error
-              ? firstError.reason.message
-              : String(firstError?.reason),
-        });
-      }
-    } else {
-      trackingConsecutiveFailures = 0;
-    }
-  });
-}
-
-async function executeSearchInternal(
-  deps: ToolDeps,
-  accessibleTables: SearchTable[],
-  query: string,
-  limit: number,
-  mode: SearchMode = "hybrid",
-  tier?: Tier,
-  offset = 0,
-  namespace?: NamespaceFilter,
-  includeLinks?: boolean,
-  sourceScope?: SourceScope,
-  options: ExecuteSearchOptions = {},
-): Promise<SearchRow[]> {
-  const enableGraph = options.enableGraph === true;
-  const ftsConfig = options.ftsConfig ?? DEFAULT_FTS_CONFIG;
-  if (sourceScope) {
-    // Neither corpus carries `source_refs`, so a source-scope predicate cannot
-    // be evaluated against them. Dropping them is the safe reading: keeping
-    // them would return rows the scope never authorized.
-    accessibleTables = accessibleTables.filter(
-      (table) => table !== "entities" && table !== "session_events",
-    );
-    if (accessibleTables.length === 0) return [];
-  }
-  let rows: SearchRow[];
-  if (mode === "keyword") {
-    rows = await ftsSearch(
-      deps,
-      accessibleTables,
-      query,
-      limit,
-      tier,
-      offset,
-      namespace,
-      sourceScope,
-      ftsConfig,
-    );
-    if (includeLinks !== false) {
-      rows = await attachExplicitLinks(deps, rows, namespace);
-    }
-    return rows;
-  }
-
-  // Vector and hybrid both need an embedding
-  const embedding = await generateSearchEmbedding(deps, query);
-  if (!embedding) {
-    // Fall back to keyword-only if embedding fails in hybrid mode
-    if (mode === "hybrid") {
-      logger.warn("embedding_failed_fallback_fts", {
-        queryLength: query.length,
-      });
-      const totalNeeded = offset + limit;
-      const fetchLimit = totalNeeded * HYBRID_FETCH_MULTIPLIER;
-      const graphRows =
-        enableGraph && sourceScope === undefined
-          ? await relationalGraphSearch(
-              deps,
-              accessibleTables,
-              query,
-              Math.min(fetchLimit, RELATIONAL_GRAPH_FETCH_LIMIT),
-              tier,
-              namespace,
-            )
-          : [];
-      if (graphRows.length > 0) {
-        const ftsRows = await ftsSearch(
-          deps,
-          accessibleTables,
-          query,
-          fetchLimit,
-          tier,
-          0,
-          namespace,
-          sourceScope,
-          ftsConfig,
-        );
-        rows = rrfMerge([], ftsRows, totalNeeded, graphRows).slice(offset);
-      } else {
-        rows = await ftsSearch(
-          deps,
-          accessibleTables,
-          query,
-          limit,
-          tier,
-          offset,
-          namespace,
-          sourceScope,
-          ftsConfig,
-        );
-      }
-      if (includeLinks !== false) {
-        rows = await attachExplicitLinks(deps, rows, namespace);
-      }
-      return rows;
-    }
-    // In vector mode, null embedding is a hard failure -- signal via thrown error
-    throw new Error("Failed to generate query embedding");
-  }
-
-  if (mode === "vector") {
-    rows = await vectorSearch(
-      deps,
-      accessibleTables,
-      embedding,
-      limit,
-      tier,
-      offset,
-      namespace,
-      sourceScope,
-    );
-    if (includeLinks !== false) {
-      rows = await attachExplicitLinks(deps, rows, namespace);
-    }
-    return rows;
-  }
-
-  // Hybrid: run both in parallel, merge with RRF
-  // Over-fetch to cover offset + limit, then slice after merge
-  const totalNeeded = offset + limit;
-  const fetchLimit = totalNeeded * HYBRID_FETCH_MULTIPLIER;
-  const [vectorRows, ftsRows] = await Promise.all([
-    vectorSearch(
-      deps,
-      accessibleTables,
-      embedding,
-      fetchLimit,
-      tier,
-      0,
-      namespace,
-      sourceScope,
-    ),
-    ftsSearch(
-      deps,
-      accessibleTables,
-      query,
-      fetchLimit,
-      tier,
-      0,
-      namespace,
-      sourceScope,
-      ftsConfig,
-    ),
-  ]);
-  const graphRows =
-    enableGraph && sourceScope === undefined
-      ? await relationalGraphSearch(
-          deps,
-          accessibleTables,
-          query,
-          Math.min(fetchLimit, RELATIONAL_GRAPH_FETCH_LIMIT),
-          tier,
-          namespace,
-        )
-      : [];
-
-  const merged = rrfMerge(vectorRows, ftsRows, totalNeeded, graphRows);
-  rows = merged.slice(offset);
-  if (includeLinks !== false) {
-    rows = await attachExplicitLinks(deps, rows, namespace);
-  }
-  return rows;
-}
-
-export function executeSearch(
-  deps: ToolDeps,
-  accessibleTables: SearchTable[],
-  query: string,
-  limit: number,
-  mode: SearchMode = "hybrid",
-  tier?: Tier,
-  offset = 0,
-  namespace?: NamespaceFilter,
-  includeLinks?: boolean,
-  sourceScope?: SourceScope,
-  options: ExecuteSearchOptions = {},
-): Promise<SearchRow[]> {
-  return traceRetrievalSpan({
-    name: "retrieval.execute",
-    input: {
-      query,
-      tables: accessibleTables,
-      limit,
-      mode,
-      tier,
-      offset,
-      namespace,
-      include_links: includeLinks,
-      source_scope: sourceScope,
-      options,
-    },
-    metadata: {
-      stage: "retrieval_pipeline",
-      resolved_namespace: namespace ?? null,
-      filter_names: [
-        "permissions",
-        "namespace",
-        "tier",
-        "source_scope",
-        "archived_at",
-        "pagination",
-      ],
-    },
-    run: () =>
-      executeSearchInternal(
-        deps,
-        accessibleTables,
-        query,
-        limit,
-        mode,
-        tier,
-        offset,
-        namespace,
-        includeLinks,
-        sourceScope,
-        options,
-      ),
-    output: rowsEvidence,
-  });
-}
-
-export async function executeSearchWithSharedFallback(
-  deps: ToolDeps,
-  accessibleTables: SearchTable[],
-  query: string,
-  limit: number,
-  mode: SearchMode,
-  tier: Tier | undefined,
-  offset: number,
-  namespace: NamespaceFilter | undefined,
-  includeLinks?: boolean,
-  sourceScope?: SourceScope,
-  options: ExecuteSearchOptions = {},
-): Promise<SearchRow[]> {
-  const config = sharedNamespaceConfig();
-  if (
-    !config.legacyFallbackEnabled ||
-    config.legacySharedNamespace === "" ||
-    offset !== 0 ||
-    namespace !== config.sharedNamespace
-  ) {
-    return withCanonicalNamespaces(
-      await executeSearch(
-        deps,
-        accessibleTables,
-        query,
-        limit,
-        mode,
-        tier,
-        offset,
-        namespace,
-        includeLinks,
-        sourceScope,
-        options,
-      ),
-    );
-  }
-
-  const sharedRows = await executeSearch(
-    deps,
-    accessibleTables,
-    query,
-    limit,
-    mode,
-    tier,
-    0,
-    config.sharedNamespace,
-    includeLinks,
-    sourceScope,
-    options,
-  );
-  if (
-    sharedRows.length >= limit ||
-    sharedRows.length >= config.fallbackMinResults
-  ) {
-    return withCanonicalNamespaces(sharedRows);
-  }
-
-  const legacyRows = await executeSearch(
-    deps,
-    accessibleTables,
-    query,
-    limit - sharedRows.length,
-    mode,
-    tier,
-    0,
-    config.legacySharedNamespace,
-    includeLinks,
-    sourceScope,
-    options,
-  );
-  return withCanonicalNamespaces(
-    mergeFallbackSearchRows(sharedRows, legacyRows, limit),
-  );
-}
-
-export async function executeSearchWithScopedSharedFallback(
-  deps: ToolDeps,
-  accessibleTables: SearchTable[],
-  query: string,
-  limit: number,
-  mode: SearchMode,
-  tier: Tier | undefined,
-  offset: number,
-  namespace: NamespaceFilter | undefined,
-  includeLinks?: boolean,
-  sourceScope?: SourceScope,
-  options: ExecuteSearchOptions = {},
-): Promise<SearchRow[]> {
-  const config = sharedNamespaceConfig();
-  const scopedNamespaces = Array.isArray(namespace) ? namespace : [];
-  if (
-    !config.legacyFallbackEnabled ||
-    config.legacySharedNamespace === "" ||
-    offset !== 0 ||
-    !scopedNamespaces.includes(config.physicalSharedNamespace)
-  ) {
-    return withCanonicalNamespaces(
-      await executeSearch(
-        deps,
-        accessibleTables,
-        query,
-        limit,
-        mode,
-        tier,
-        offset,
-        namespace,
-        includeLinks,
-        sourceScope,
-        options,
-      ),
-    );
-  }
-
-  const [primaryRows, sharedRows] = await Promise.all([
-    executeSearch(
-      deps,
-      accessibleTables,
-      query,
-      limit,
-      mode,
-      tier,
-      0,
-      namespace,
-      includeLinks,
-      sourceScope,
-      options,
-    ),
-    executeSearch(
-      deps,
-      accessibleTables,
-      query,
-      limit,
-      mode,
-      tier,
-      0,
-      config.physicalSharedNamespace,
-      includeLinks,
-      sourceScope,
-      options,
-    ),
-  ]);
-  if (
-    sharedRows.length >= limit ||
-    sharedRows.length >= config.fallbackMinResults
-  ) {
-    return withCanonicalNamespaces(primaryRows);
-  }
-
-  const legacyRows = await executeSearch(
-    deps,
-    accessibleTables,
-    query,
-    limit,
-    mode,
-    tier,
-    0,
-    config.legacySharedNamespace,
-    includeLinks,
-    sourceScope,
-    options,
-  );
-  return withCanonicalNamespaces(
-    mergeFallbackSearchRows(primaryRows, legacyRows, limit),
-  );
-}
-
-export function registerSearchBrain(server: McpServer, deps: ToolDeps): void {
-  server.registerTool(
-    "search_brain",
-    {
-      description:
-        "Search across all brain tables. Supports hybrid (vector + keyword), pure vector, or keyword-only modes.",
-      inputSchema: {
-        query: z.string().min(1).describe("Natural language search query"),
-        table: z
-          .enum([
-            "thoughts",
-            "decisions",
-            "relationships",
-            "projects",
-            "sessions",
-            "entities",
-            "session_events",
-          ])
-          .optional()
-          .describe("Optional: restrict search to a single named source"),
-        namespace: z
-          .string()
-          .trim()
-          .min(1)
-          .max(500)
-          .optional()
-          .describe(
-            "Optional: filter results to a specific namespace (e.g. clientId or 'shared-kb')",
-          ),
-        limit: z
-          .number()
-          .int()
-          .min(1)
-          .max(250)
-          .optional()
-          .describe("Maximum results to return (default 10)"),
-        offset: z
-          .number()
-          .int()
-          .min(0)
-          .optional()
-          .describe("Number of results to skip for pagination (default 0)"),
-        search_mode: z
-          .enum(["hybrid", "vector", "keyword"])
-          .optional()
-          .describe(
-            "Search mode: hybrid (default) = vector + keyword with RRF fusion, vector = semantic only, keyword = full-text only",
-          ),
-        tier: z
-          .enum(["hot", "warm", "cold"])
-          .optional()
-          .describe("Optional: filter results to a specific cognitive tier"),
-        source_scope: sourceScopeSchema
-          .optional()
-          .describe(
-            "Optional: require matching source reference client_id, matter_id, document_id, path, and/or dms_id.",
-          ),
-        fts_config: z
-          .string()
-          .trim()
-          .min(1)
-          .max(64)
-          .optional()
-          .describe(
-            `Optional: keyword full-text-search language configuration for this request. ` +
-              `Accepts a supported Postgres regconfig (${SUPPORTED_FTS_CONFIGS.join(", ")}) ` +
-              `or a language token (e.g. 'de', 'de-DE', 'spanish'). Unrecognized values ` +
-              `fall back to the deployment corpus default (OPENBRAIN_FTS_CONFIG, else english). ` +
-              `An explicitly requested effective non-English config requires admin or ob-admin ` +
-              `for keyword/hybrid searches; vector mode performs no FTS, so fts_config is ` +
-              `ignored there. For ordinary roles, a non-English deployment env default ` +
-              `degrades to english rather than denying. ` +
-              `Affects keyword/hybrid stemming only; english is byte-identical to prior behavior.`,
-          ),
-      },
-      annotations: {
-        title: "Search Brain",
-        readOnlyHint: true,
-        destructiveHint: false,
-        idempotentHint: true,
-      },
-    },
-    async (args, extra) => {
-      const auth = extra.authInfo as AuthInfo | undefined;
-      if (!auth) {
-        return {
-          content: [
-            {
-              type: "text" as const,
-              text: "Permission denied: no readable tables",
-            },
-          ],
-          isError: true,
-        };
-      }
-
-      const tableFilter = args.table as SearchTable | undefined;
-      let accessibleTables: SearchTable[];
-
-      if (tableFilter) {
-        if (tableFilter === "entities") {
-          if (!canRead(auth.role, "sessions")) {
-            return {
-              content: [
-                {
-                  type: "text" as const,
-                  text: "Permission denied: cannot read entities",
-                },
-              ],
-              isError: true,
-            };
-          }
-          accessibleTables = ["entities"];
-        } else if (tableFilter === "session_events") {
-          // Session events are session-scoped content, so they ride the
-          // `sessions` read permission -- the same indirection `entities` uses
-          // above. There is no `session_events` row in PERMISSIONS because the
-          // corpus is read-only through this surface and never written here.
-          if (!canRead(auth.role, "sessions")) {
-            return {
-              content: [
-                {
-                  type: "text" as const,
-                  text: "Permission denied: cannot read session_events",
-                },
-              ],
-              isError: true,
-            };
-          }
-          accessibleTables = ["session_events"];
-        } else if (!canRead(auth.role, tableFilter)) {
-          return {
-            content: [
-              {
-                type: "text" as const,
-                text: `Permission denied: cannot read ${tableFilter}`,
-              },
-            ],
-            isError: true,
-          };
-        } else {
-          accessibleTables = [tableFilter];
-        }
-      } else {
-        accessibleTables = readableSearchTables(auth.role, {
-          includeEntities: true,
-        });
-      }
-
-      if (accessibleTables.length === 0) {
-        return {
-          content: [
-            {
-              type: "text" as const,
-              text: "Permission denied: no readable tables",
-            },
-          ],
-          isError: true,
-        };
-      }
-
-      const limit = args.limit ?? 10;
-      const offset = args.offset ?? 0;
-      const mode = (args.search_mode as SearchMode) ?? "hybrid";
-      const tier = args.tier as Tier | undefined;
-      const requestedNamespace = args.namespace as string | undefined;
-      const sourceScope = args.source_scope as SourceScope | undefined;
-      const requestedFtsConfig = args.fts_config as string | undefined;
-      const ftsPrivileged = auth.role === "admin" || auth.role === "ob-admin";
-      // The non-English privilege boundary applies to the EFFECTIVE config for
-      // keyword/hybrid searches regardless of provenance (request argument or
-      // OPENBRAIN_FTS_CONFIG env default), because either one selects the
-      // unindexed on-the-fly to_tsvector path (#368 post-merge finding 1).
-      let ftsConfig = requestFtsConfig(requestedFtsConfig);
-      if (ftsConfig !== DEFAULT_FTS_CONFIG && !ftsPrivileged) {
-        if (mode !== "vector" && requestedFtsConfig !== undefined) {
-          // An ordinary role explicitly asked for an effective non-English
-          // config on a mode that runs FTS: content-free denial (unchanged).
-          return {
-            content: [
-              {
-                type: "text" as const,
-                text: NON_ENGLISH_FTS_AUTHORIZATION_ERROR,
-              },
-            ],
-            isError: true,
-          };
-        }
-        // Non-English arrived only via the operator env default, or the mode
-        // is vector (which performs no FTS, so fts_config is unused -- #368
-        // post-merge finding 2). Degrade to the GIN-indexed english default
-        // instead of denying: ordinary roles keep exactly the pre-#341
-        // availability and cost profile, and an unused argument can neither
-        // deny nor influence execution.
-        ftsConfig = DEFAULT_FTS_CONFIG;
-      }
-      const sourceScopeError = sourceScopeAuthorizationError(auth, sourceScope);
-      if (sourceScopeError) {
-        return {
-          content: [{ type: "text" as const, text: sourceScopeError }],
-          isError: true,
-        };
-      }
-      if (sourceScope) {
-        // Same reason as executeSearchInternal: no source_refs column means a
-        // source scope cannot be honored, so the corpus is excluded rather
-        // than silently ignoring the scope.
-        accessibleTables = accessibleTables.filter(
-          (table) => table !== "entities" && table !== "session_events",
-        );
-      }
-      if (accessibleTables.length === 0) {
-        return {
-          content: [
-            {
-              type: "text" as const,
-              text: "No source-scoped tables are readable",
-            },
-          ],
-        };
-      }
-      if (requestedNamespace && !canReadNamespace(auth, requestedNamespace)) {
-        return {
-          content: [
-            {
-              type: "text" as const,
-              text: "Permission denied: namespace read access denied",
-            },
-          ],
-          isError: true,
-        };
-      }
-      const namespace = namespaceFilterFor(auth, requestedNamespace);
-      setActiveMcpTraceMetadata({ resolved_namespace: namespace ?? null });
-      const shouldUseSharedFallback =
-        requestedNamespace !== undefined &&
-        isSharedNamespace(requestedNamespace);
-
-      let rows;
-      try {
-        rows = shouldUseSharedFallback
-          ? await executeSearchWithSharedFallback(
-              deps,
-              accessibleTables,
-              args.query,
-              limit,
-              mode,
-              tier,
-              offset,
-              namespace,
-              undefined,
-              sourceScope,
-              { enableGraph: true, ftsConfig },
-            )
-          : await executeSearchWithScopedSharedFallback(
-              deps,
-              accessibleTables,
-              args.query,
-              limit,
-              mode,
-              tier,
-              offset,
-              namespace,
-              undefined,
-              sourceScope,
-              { enableGraph: true, ftsConfig },
-            );
-      } catch (err) {
-        // "An empty result set" is this repo's named anti-pattern
-        // (docs/GOTCHAS.md): a failed search and a search that found nothing
-        // were indistinguishable in the logs. The pg fields exist only here.
-        logger.error("search_brain_failed", {
-          namespace,
-          mode,
-          tier,
-          ...describeError(err),
-        });
-        const message = err instanceof Error ? err.message : String(err);
-        return {
-          content: [{ type: "text" as const, text: message }],
-          isError: true,
-        };
-      }
-
-      trackUsage(deps, rows, args.query, "search", auth.clientId);
-
-      return {
-        content: [
-          {
-            type: "text" as const,
-            text: JSON.stringify(rows),
-          },
-        ],
-      };
-    },
-  );
+export function trackUsage(...args: TrackUsageArgs): void {
+  const [deps, rows, queryText, context = "search", accessedBy] = args;
+  trackUsageOptions({ deps, rows, queryText, context, accessedBy });
 }
